fix test in windows
